+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 (Add entries here in the order of: server, console, cli, docs, others)
 
+- server: add 'replace_configuration' option (default: false) in the add source API payload
 - server: add a comment field for actions (#231)
 - server: accept GeoJSON for MSSQL geometry and geography operators (#787)
 - console: add a comment field for actions (#231)